pyproject.toml
[tool.poetry]
name = "VPR-data_landing"
version = "0.1.0"
description = "A ETL pipeline that reads preprocessed JSON files from S3, transforms them using PySpark, and writes them back to S3."
authors = ["Miguel Granica <migue.granica@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34"
pyspark = "^3.5.1"
pyyaml = "^6.0"
delta-spark = "^3.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.dev-dependencies]
pytest = "^8.3"  # For testing
pytest-cov = "^4.1"  # For test coverage
black = "^23.7"  # Code formatting
isort = "^5.13"  # Import sorting
flake8 = "^6.1"  # Linting
jupyterlab = "^4.2.4"  # JupyterLab for interactive development
findspark = "^2.0"  # Utility to help find Spark home in JupyterLab

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88